Takes image file and outputs 9 cropped images.
Takes image file (image.ext), crops 9 different ways and saves each to a new file with a similar name and same file type (image1.ext, etc.). The idea is that you take a picture, run it through this program and then upload the 9 new images to instagram. Note that each new image is 3 times smaller than the original, so instagram will blow it up to the required size; so make sure your image is high res!